+++ npm/build-wiki.js
@@ -1,25 +1,42 @@
 #!/usr/bin/env node
+/* eslint-env node, es6 */
 // ---------------------------------------------------------------------------------------------------------------------
 // This script is intended to generate a wiki for this module
 // ---------------------------------------------------------------------------------------------------------------------
 
 require('shelljs/global');
-require('colors');
 
-// stop on encountering the first error
-set('-e');
-echo('-e', 'Generating wiki...');
-echo('-e', 'jsdoc2md');
+var fs = require('fs'),
 
-// some variables
-var OUT_DIR = 'out/wiki',
+    colors = require('colors/safe'),
+    jsdoc2md = require('jsdoc-to-markdown'),
+
+    OUT_DIR = 'out/wiki',
+    INP_DIR = 'lib/**/*.js',
     OUT_FILE = 'REFERENCE.md',
-    OUT_PATH = `${OUT_DIR}/${OUT_FILE}`;
+    OUT_PATH = OUT_DIR + '/' + OUT_FILE,
+    SUCCESS_MESSAGE = colors.green.bold(`- wiki generated at "${OUT_PATH}"`);
 
-// clean directory
-test('-d', OUT_DIR) && rm('-rf', OUT_DIR);
-mkdir('-p', OUT_DIR);
+module.exports = function (exit) {
+    console.info(colors.yellow.bold('Generating wiki using jsdoc2md...'));
 
-// execute command
-exec(`jsdoc2md --src lib/**/*.js > ${OUT_PATH};`);
-echo(` - wiki generated at ${OUT_PATH}\n`);
+    // clean directory
+    test('-d', OUT_DIR) && rm('-rf', OUT_DIR);
+    mkdir('-p', OUT_DIR);
+
+    // execute command
+    jsdoc2md.render({ files: INP_DIR })
+        .then(function (markdown) {
+            fs.writeFile(OUT_PATH, markdown, function (err) {
+                console.info(err ? err : SUCCESS_MESSAGE);
+                exit(err ? 1 : 0);
+            });
+        })
+        .catch(function (err) {
+            console.error(err);
+            exit(1);
+        });
+};
+
+// ensure we run this script exports if this is a direct stdin.tty run
+!module.parent && module.exports(exit);